#659 move NvBufSurfaceGenerator to a separate lib
