+++ libs/nvbufsurfacegenerator/requirements.txt
@@ -1,0 +1,3 @@
+setuptools>=61.0
+pybind11[global]>=2.11
+scikit-build>=0.17
